fix(precision): keep AGP/SD geminal in fp64 to prevent fp32 amplification of log|det|
Diagnostic scripts in bug/fp32/ showed that with the previous mixed-precision
defaults (orb_eval=fp32, geminal=fp32), local energies on a 32-up/32-dn ECP
system were biased vs full fp64. The bias came from the kinetic term: AO matrix
entries had only fp32 round-off, but the determinant amplified that into a log|det|
error, which fed back into both T and ECP non-local matrix elements.

Two minimal changes restore safety while keeping the heavy AO kernels in fp32:

1. molecular_orbital.compute_MOs now upcasts the (small) MO matmul to the
   determinant-zone dtype (fp64 in mixed mode). The expensive AO evaluation
   still runs in orb_eval (fp32), but the MO matrix returned to the
   determinant / kinetic / energy paths is fp64, breaking the chain
   "fp32 AO -> fp32 geminal -> noisy log|det|".

2. _precision._DEFAULTS_MIXED["geminal"] is changed from float32 to float64.
   The geminal matrix is the input to LU/det; even ~1e-7 entry noise blows
   up to O(1) errors in log|det|. The diag_04 AGP sweep additionally
   shows that AO-direct (AGP) form is ~4x more sensitive to geminal=fp32
   than the MO (JSD) form so this default is required for both ansatz types.

+++ jqmc/_precision.py
@@ -111,9 +111,14 @@ _DEFAULTS_FULL: dict[str, str] = {
 
 # --- mode="mixed" defaults (recommended mixed precision) ---
 # float32 risk:
-#   orb_eval    - low: smooth Gaussian basis + linear combination
+#   orb_eval    - low: smooth Gaussian basis + linear combination.
+#                 (Heavy AO eval stays in fp32; compute_MOs upcasts the small
+#                  matmul to the determinant zone, see molecular_orbital.py.)
 #   jastrow     - low: smooth correlation function, pre-exp value
-#   geminal     - low: building matrix elements only (pre-det)
+#   geminal     - HIGH: this matrix is the input to LU/det; even ε≈1e-7 on
+#                 entries amplifies into log|det| errors of O(1) for ~32x32
+#                 systems with non-trivial condition numbers
+#                 (see bug/fp32 diagnostics).  Kept in fp64.
 #   coulomb     - low-medium: sum of 1/r + ECP spherical quadrature
 #   determinant - high: log(det) cancellation, SVD 1/s, eigenvalue ops
 #   kinetic     - high: second derivative of ln|Psi|, cancellation-sensitive
@@ -122,9 +127,9 @@ _DEFAULTS_FULL: dict[str, str] = {
 #   optimization- high: S^{-1}F linear system, ill-conditioned matrix
 #   io          - low-medium: file I/O + nuclear coordinates
 _DEFAULTS_MIXED: dict[str, str] = {
-    "orb_eval": "float32",  # low risk
+    "orb_eval": "float32",  # low risk (heavy kernel only)
     "jastrow": "float32",  # low risk
-    "geminal": "float32",  # low risk
+    "geminal": "float64",  # high risk: feeds LU/det
     "determinant": "float64",  # high risk
     "coulomb": "float32",  # low-medium risk
     "kinetic": "float64",  # high risk
